Do not show user API keys
This updates the Administration > General view of the Dashboard to not
show the API key of the user that connects with the Storage
Service. The current plain text field has been replaced with a
password field to do so.

Also the Dashboard user views and their redirects have been updated to
show the API key only once on re-generation. This uses Django's
messages framework to render an alert with a button to copy the API
key to the browser's clipboard.