Add libreoffice to Docker image
We want to be able to convert various word documents to PDFs. Headless
libreoffice seems to be the best tool for doing this, so this commit
adds it to our docker images. This allows us to add normalization rules
using libreoffice to our Archivematica instance manually via the
dashboard
(https://archivematica.org/en/docs/archivematica-1.17/user-manual/preservation/preservation-planning/#altering-commands-rules)