Optimize dashboard queries in status endpoints
Currently, the queries used to retrieve items for the Transfer and
Ingest tabs on the dashboard first retrieve all possible items, then go
through them  one by one and weed out the ones that are supposed to be
hidden. This commit updates the transfer and ingest status endpoints to
ignore hidden items on the first pass instead. In cases where there are
many hidden items, this avoids numerous unnecessary database calls and
speeds up the transfer/status and ingest/status endpoints significantly.
The commit also adds tests for the RPCServer's _units_statuses_handler.